Added Google Sign-In with Firebase Authentication
This commit introduces the capability to sign in with Google using Firebase Authentication in the login.jsx and registration.jsx files. With this enhancement, users can now conveniently use their Google accounts to log in or register, enhancing the authentication options available in our application. The integration leverages Firebase's robust authentication system to provide a seamless and secure sign-in experience for our users, improving overall usability and accessibility.