Updated Client-Side Website for Firebase Authentication
In this commit, I've made significant updates to our client-side website, specifically focusing on user authentication using Firebase Authentication. This enhancement marks a pivotal step in our application's development, as it empowers us to manage user authentication seamlessly, paving the way for dynamic dashboards and personalised experiences for our users.

Key Changes and Enhancements:

Authentication State Observer: Implemented an authentication state observer to monitor and track the user's authentication status in real-time. This observer ensures that we always have the latest information regarding whether a user is signed in or not.

Personalized Dashboards: With this foundation in place, we are now well-equipped to create dynamic dashboards and provide personalized information and experiences tailored to each user. This is a significant milestone as it opens up opportunities for delivering user-specific content and functionality.

These changes not only enhance the security and reliability of our authentication system but also set the stage for us to build user-centric features and capabilities that will drive engagement and satisfaction.