Enhanced User Profile and Patient Details Pages
In this commit, I made significant improvements to the user experience by enhancing the Profile and PatientDetail pages. The key changes are as follows:

Tabular Information Display: The Profile and PatientDetail pages have been updated to display user information in a tabular format. This provides a structured and organised view of the user's details, making it more user-friendly.

Edit Mode: Users can now easily edit their information. In the Profile page, I introduced an edit mode that enables users to update their personal, home, and work details conveniently. This enhances the user's ability to maintain their information accurately.

Stylish Table Layout: The table layout of the information display has been designed to be visually appealing, ensuring a consistent and aesthetically pleasing user interface.

Webhook Handler for Stripe: I also updated the webhook handler to listen for the charge.succeeded event from Stripe. This enhancement allows us to capture and save the last 4 digits of a card used in a transaction to our database dynamically.