Fixed Filtering and Duplicate Attribute Bugs in Patient Information
In this update, we've addressed two critical bugs that were affecting the functionality of our patient information system. These bug fixes enhance the user experience and ensure data integrity in our application.

Bugs Fixed:

Filter for Male or Female: Previously, the filter for male or female patients was case-sensitive, which led to filtering issues. We've corrected this behaviour by updating the filter options to "Male" and "Female".

Duplicate Attribute Issue: A critical bug that caused attributes to be duplicated, with some returning "NaN," has been resolved. We identified the root cause and implemented a fix by converting the numericFields in patientDetails to lowercase, preventing attribute duplication and data corruption.