Add Fields For Heart Prediction & Pages For Prediction Results
This commit extends the functionality of the healthcare application by incorporating fields to capture crucial information essential for predicting heart conditions. Furthermore, distinct pages for lung, heart, and breast cancer prediction results have been introduced to enhance user experience and streamline the prediction process.

A key feature of this update is the implementation of an intelligent attribute retrieval system. Now, when users click on a prediction button, the application dynamically gathers the relevant patient attributes. This ensures a seamless flow of information, preparing the groundwork for the integration of a machine learning model's server for accurate predictions in the future.